HSEARCH-4620 Move new IndexingPlan-related classes to the mapper.pojo.work package
