+++ integrationtest/mapper/pojo-base/src/test/java/org/hibernate/search/integrationtest/mapper/pojo/work/PojoIndexingPlanSynchronizationStrategyIT.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.integrationtest.mapper.pojo.plan.synchronization;
+package org.hibernate.search.integrationtest.mapper.pojo.work;
 
 import java.lang.invoke.MethodHandles;
 
@@ -13,7 +13,7 @@ import org.hibernate.search.engine.backend.work.execution.DocumentRefreshStrateg
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.DocumentId;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.GenericField;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.Indexed;
-import org.hibernate.search.mapper.pojo.plan.synchronization.IndexingPlanSynchronizationStrategy;
+import org.hibernate.search.mapper.pojo.work.IndexingPlanSynchronizationStrategy;
 import org.hibernate.search.mapper.pojo.standalone.mapping.SearchMapping;
 import org.hibernate.search.mapper.pojo.standalone.session.SearchSession;
 import org.hibernate.search.util.impl.integrationtest.common.rule.BackendMock;
@@ -23,7 +23,7 @@ import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 
-public class IndexingPlanSynchronizationStrategyIT {
+public class PojoIndexingPlanSynchronizationStrategyIT {
 
 	@Rule
 	public BackendMock backendMock = new BackendMock();
@@ -59,6 +59,7 @@ public class IndexingPlanSynchronizationStrategyIT {
 					);
 		}
 
+		// Note all kinds of overrides are already tested in depth in AbstractPojoIndexingOperationIT and subclasses
 		try ( SearchSession session = mapping.createSession() ) {
 			session.indexingPlanSynchronizationStrategy( IndexingPlanSynchronizationStrategy.sync() );
