+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/automaticindexing/session/AutomaticIndexingSynchronizationStrategyNames.java
@@ -7,7 +7,7 @@
 package org.hibernate.search.mapper.orm.automaticindexing.session;
 
 import org.hibernate.search.mapper.orm.cfg.HibernateOrmMapperSettings;
-import org.hibernate.search.mapper.pojo.plan.synchronization.IndexingPlanSynchronizationStrategyNames;
+import org.hibernate.search.mapper.pojo.work.IndexingPlanSynchronizationStrategyNames;
 
 /**
  * The names of built-in automatic indexing synchronization strategies,
