+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/session/impl/AfterCommitIndexingPlanSynchronization.java
@@ -13,7 +13,7 @@ import javax.transaction.Synchronization;
 import org.hibernate.Transaction;
 import org.hibernate.search.mapper.orm.common.EntityReference;
 import org.hibernate.search.mapper.orm.logging.impl.Log;
-import org.hibernate.search.mapper.pojo.plan.synchronization.spi.ConfiguredIndexingPlanSynchronizationStrategy;
+import org.hibernate.search.mapper.pojo.work.spi.ConfiguredIndexingPlanSynchronizationStrategy;
 import org.hibernate.search.mapper.pojo.work.spi.PojoIndexingPlan;
 import org.hibernate.search.util.common.logging.impl.LoggerFactory;
