+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/work/impl/SearchIndexingPlanSessionContext.java
@@ -8,7 +8,7 @@ package org.hibernate.search.mapper.orm.work.impl;
 
 import org.hibernate.search.mapper.orm.common.EntityReference;
 import org.hibernate.search.mapper.pojo.model.spi.PojoRuntimeIntrospector;
-import org.hibernate.search.mapper.pojo.plan.synchronization.spi.ConfiguredIndexingPlanSynchronizationStrategy;
+import org.hibernate.search.mapper.pojo.work.spi.ConfiguredIndexingPlanSynchronizationStrategy;
 import org.hibernate.search.mapper.pojo.work.spi.PojoIndexingPlan;
 
 public interface SearchIndexingPlanSessionContext {
