+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/impl/PojoBaseBeanConfigurer.java
@@ -9,8 +9,8 @@ package org.hibernate.search.mapper.pojo.impl;
 import org.hibernate.search.engine.environment.bean.BeanReference;
 import org.hibernate.search.engine.environment.bean.spi.BeanConfigurationContext;
 import org.hibernate.search.engine.environment.bean.spi.BeanConfigurer;
-import org.hibernate.search.mapper.pojo.plan.synchronization.IndexingPlanSynchronizationStrategy;
-import org.hibernate.search.mapper.pojo.plan.synchronization.IndexingPlanSynchronizationStrategyNames;
+import org.hibernate.search.mapper.pojo.work.IndexingPlanSynchronizationStrategy;
+import org.hibernate.search.mapper.pojo.work.IndexingPlanSynchronizationStrategyNames;
 
 public class PojoBaseBeanConfigurer implements BeanConfigurer {
 	@Override
