+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/work/IndexingPlanSynchronizationStrategyConfigurationContext.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.pojo.plan.synchronization;
+package org.hibernate.search.mapper.pojo.work;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
@@ -13,7 +13,6 @@ import org.hibernate.search.engine.backend.work.execution.DocumentCommitStrategy
 import org.hibernate.search.engine.backend.work.execution.DocumentRefreshStrategy;
 import org.hibernate.search.engine.backend.work.execution.OperationSubmitter;
 import org.hibernate.search.engine.reporting.FailureHandler;
-import org.hibernate.search.mapper.pojo.work.SearchIndexingPlanExecutionReport;
 import org.hibernate.search.util.common.annotation.Incubating;
 
 @Incubating
